Document_Parsing/requirements.txt
chainlit
google-generativeai
pytesseract
pillow
python-docx
PyPDF2
